requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
pydantic==2.11.3
langchain==0.3.24
langchain_groq==0.3.2
mcp-use==1.2.8
